proc_spy: expose to library users
