+++ lib/mpv.rb
@@ -10,6 +10,7 @@ require_relative "ass/color"
 require_relative "ass/text"
 
 require_relative "mpv/exceptions"
+require_relative "mpv/proc_spy"
 require_relative "mpv/multi_queue"
 require_relative "mpv/utils"
 require_relative "mpv/client"
